MS VS2012 fixes preparing for v2.5 release.

+++ dev/ext/lib/Geomatics/SolarSystem.cpp
@@ -1051,7 +1051,7 @@ try {
    T = 2.0*(tt-Tbeg)/Tspan - 1.0;
 
    // interpolate
-   unsigned int N=c_ncoeff[which];
+   long N=c_ncoeff[which];
    vector<double> C(N,0.0);     // Chebyshev
    vector<double> U(N,0.0);     // derivative of Chebyshev
    for(i=0; i<ncomp; i++) {     // loop over components
